// pkg/packages/azure/src/generated/container-apps/model/containerAppCollection.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * ContainerApps API Client
 * OpenAPI spec version: 2025-07-01
 */
import type { ContainerApp } from "./containerApp";

/**
 * Container App collection ARM resource.
 */
export interface ContainerAppCollection {
  /** Collection of resources. */
  value: ContainerApp[];
  /** Link to next page of resources. */
  readonly nextLink?: string;
}