// pkg/packages/azure/src/generated/container-apps/model/CommonDefinitions/volume.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * ContainerApps API Client
 * OpenAPI spec version: 2025-07-01
 */
import type { VolumeStorageType } from './volumeStorageType';
import type { SecretVolumeItem } from './secretVolumeItem';

/**
 * Volume definitions for the Container App.
 */
export interface Volume {
  /** Volume name. */
  name?: string;
  /** Storage type for the volume. If not provided, use EmptyDir. */
  storageType?: VolumeStorageType;
  /** Name of storage resource. No need to provide for EmptyDir and Secret. */
  storageName?: string;
  /** List of secrets to be added in volume. If no secrets are provided, all secrets in collection will be added to volume. */
  secrets?: SecretVolumeItem[];
  /** Mount options used while mounting the Azure file share or NFS Azure file share. Must be a comma-separated string. */
  mountOptions?: string;
}